made fixes
